kludge: dust off the feat-non-generic branch from 2020
Because of a change in how Cargo normalizes dependency URLs, we are now
unable to change the `ut-utp/core` branch that this crate depends On
from a dependent. Our old lock files from May of 2020 were also no help
because they referred to branches that have since been merged and
deleted.

So, just to get unstuck while we work on newer more robust peripheral
implementations, we're taking the `feat-non-generic` branch as it was
in May of 2020 and changing it to use the latest `ut-utp/core` from
`master`.

See: https://github.com/rust-lang/cargo/issues/5478#issuecomment-1055365283